split bambu, chisel, sv2v
