+++ siliconcompiler/tools/sv2v/convert.py
@@ -1,0 +1,35 @@
+
+def setup(chip):
+    ''' Per tool function that returns a dynamic options string based on
+    the dictionary settings.
+    '''
+
+    chip.logger.debug("Setting up sv2v")
+
+    tool = 'sv2v'
+    step = chip.get('arg','step')
+    index = chip.get('arg','index')
+    task = chip.get_task(step, index)
+
+    chip.set('tool', tool, 'exe', tool)
+    chip.set('tool', tool, 'vswitch', '--numeric-version')
+    chip.set('tool', tool, 'version', '>=0.0.9', clobber=False)
+
+    chip.set('tool', tool, 'task', task, 'threads', step, index,  4, clobber=False)
+
+    # Since we run sv2v after the import/preprocess step, there should be no
+    # need for specifying include dirs/defines. However we don't want to pass
+    # --skip-preprocessor because there may still be unused preprocessor
+    # directives not removed by the importer and passing the --skip-preprocessor
+    # flag would cause sv2v to error.
+
+    # since this step should run after import, the top design module should be
+    # set and we can read the pickled Verilog without accessing the original
+    # sources
+    topmodule = chip.top()
+    chip.set('tool', tool, 'task', task, 'option', step, index,  [])
+    chip.add('tool', tool, 'task', task, 'option', step, index,  "inputs/" + topmodule + ".v")
+    chip.add('tool', tool, 'task', task, 'option', step, index,  "--write=outputs/" + topmodule + ".v")
+
+    chip.set('tool', tool, 'task', task, 'input', step, index, f'{topmodule}.v')
+    chip.set('tool', tool, 'task', task, 'output', step, index, f'{topmodule}.v')
